refactor(exchange): rename src/exchange/ folder to src/broker/
Aligns the folder name with the abstraction it contains. The npm
package name (@typedtrader/exchange) is intentionally left alone —
that's a registry-level rename and out of scope here.

- git mv preserves history for every file in the tree.
- Internal import paths (./exchange/, ../exchange/, ../../exchange/)
  rewritten to ./broker/, etc.
- src/index.ts barrel re-exports updated.
- External consumers are unaffected: they import from
  '@typedtrader/exchange', not from internal paths.
