+++ packages/exchange/src/broker/alpaca/api/schema/BarSchema.ts
