+++ packages/exchange/src/broker/alpaca/api/schema/SnapshotSchema.ts
