+++ packages/exchange/src/broker/alpaca/api/schema/TradingStreamSchema.ts
