+++ packages/exchange/src/candle/CandleBatcher.test.ts
@@ -13,7 +13,7 @@ import one_day_in_minutes from '../../fixtures/candles/candle-batcher/one_day_in
 import one_hour_in_minutes from '../../fixtures/candles/candle-batcher/one_hour_in_minutes.json' with {type: 'json'};
 import {BatchedCandle} from './BatchedCandle.js';
 import {CandleBatcher} from './CandleBatcher.js';
-import {Candle} from '../exchange/Broker.js';
+import {Candle} from '../broker/Broker.js';
 
 describe('CandleBatcher', () => {
   describe('amountOfCandles', () => {
