+++ packages/exchange/src/index.ts
@@ -1,6 +1,6 @@
 export * from './candle/index.js';
-export * from './exchange/index.js';
-export * from './exchange/alpaca/index.js';
-export * from './exchange/trading212/index.js';
+export * from './broker/index.js';
+export * from './broker/alpaca/index.js';
+export * from './broker/trading212/index.js';
 export * from './trader/index.js';
 export * from './util/index.js';
